+++ LJMD_serial/tests/test_main.cpp
@@ -1,0 +1,8 @@
+#include "gtest/gtest.h"
+#include "prototypes.h"
+#include "structures.h"
+
+int main(int argc, char **argv) {
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
